fix(deps): replace react-json-diff by monaco editor to display config
