+++ package.json
@@ -89,7 +89,6 @@
     "react-dom": "18.1.0",
     "react-flow-renderer": "10.2.2",
     "react-icons": "4.3.1",
-    "react-json-view": "1.21.3",
     "react-router": "6.3.0",
     "react-router-dom": "6.3.0",
     "reactstrap": "9.0.2",
