+++ src/client/Activation/components/ConfigJsonRenderer.jsx
@@ -1,41 +1,50 @@
 import React from 'react'
 import { Spinner } from 'reactstrap'
-import ReactJson from 'react-json-view'
 import { PropTypes } from 'prop-types'
+import MonacoEditor from '@monaco-editor/react'
 
-const ConfigJsonRenderer = ({ loading, activeConfig, newConfig, isModified }) => (
-  <>
-    {loading ? (
-      <div className="spinner-container">
-        <Spinner color="primary" />
-      </div>
-    ) : null}
-    {activeConfig && (
-    <ReactJson
-      src={activeConfig}
-      name="Active configuration"
-      collapsed
-      displayObjectSize={false}
-      displayDataTypes={false}
-      enableClipboard
-      collapseStringsAfterLength={100}
-      style={{ marginBottom: '15px' }}
-    />
-    )}
-    {newConfig && isModified && (
-    <ReactJson
-      src={newConfig}
-      name="New configuration"
-      collapsed
-      displayObjectSize={false}
-      displayDataTypes={false}
-      enableClipboard
-      collapseStringsAfterLength={100}
-      style={{ marginBottom: '15px' }}
-    />
-    )}
-  </>
-)
+const ConfigJsonRenderer = ({ loading, activeConfig, newConfig, isModified }) => {
+  const monacoEditorOptions = {
+    selectOnLineNumbers: true,
+    minimap: { enabled: false },
+    scrollbar: { alwaysConsumeMouseWheel: false },
+    readOnly: true,
+  }
+
+  return (
+    <>
+      {loading ? (
+        <div className="spinner-container">
+          <Spinner color="primary" />
+        </div>
+      ) : null}
+      {activeConfig && !loading && (
+        <div>
+          <h5>Active configuration</h5>
+          <MonacoEditor
+            language="javascript"
+            theme="vs"
+            height="400px"
+            value={JSON.stringify(activeConfig, null, 2)}
+            options={monacoEditorOptions}
+          />
+        </div>
+      )}
+      {newConfig && isModified && !loading && (
+        <div>
+          <h5>New configuration</h5>
+          <MonacoEditor
+            language="json"
+            theme="vs"
+            height="400px"
+            value={JSON.stringify(newConfig, null, 2)}
+            options={monacoEditorOptions}
+          />
+        </div>
+      )}
+    </>
+  )
+}
 
 ConfigJsonRenderer.propTypes = {
   loading: PropTypes.bool,
